keep ProtectedResourceMetadata exported without experiment tag
